Recover REPL between invocations

+++ src/modules/io.c
@@ -160,7 +160,7 @@ static Value readLineNative(VM* vm, int argc, Value* argv) {
 
     char* line = NULL;
     size_t cap = 0;
-    size_t len = getline(&line, &cap, file->file);
+    int len = getline(&line, &cap, file->file);
 
     if (len == -1) {
         free(line);
